Added support for macOS, tvOS and watchOS. Bumped the PromiseKit dependency version. Replaced the previous project with another one which includes framework and test targets for all the supported platforms. Updated the Travis CI script.
