--- Example/Podfile
@@ -1,16 +1,0 @@
-platform :ios, '9.0'
-
-source 'https://github.com/CocoaPods/Specs.git'
-use_frameworks!
-
-def shared_pods
-  pod 'PromiseKit', '~> 4'
-end
-
-target 'AwaitKitExample' do
-  shared_pods
-end
-
-target 'AwaitKitTests' do
-  shared_pods
-end
