+++ Podfile
@@ -1,0 +1,36 @@
+abstract_target 'Framework' do
+  use_frameworks!
+  pod 'PromiseKit', '~> 4.1'
+
+  target 'AwaitKit iOS' do
+    platform :ios, '8.0'
+  end
+
+  target 'AwaitKit iOS Tests' do
+    platform :ios, '8.0'
+  end
+
+  target 'AwaitKit macOS' do
+    platform :osx, '10.10'
+  end
+
+  target 'AwaitKit macOS Tests' do
+    platform :osx, '10.10'
+  end
+
+  target 'AwaitKit tvOS' do
+    platform :tvos, '9.0'
+  end
+
+  target 'AwaitKit tvOS Tests' do
+    platform :tvos, '9.0'
+  end
+
+  target 'AwaitKit watchOS' do
+    platform :watchos, '2.0'
+  end
+
+  target 'Example' do
+    platform :ios, '8.0'
+  end
+end
