+++ Sources/AwaitKit.h
@@ -1,0 +1,19 @@
+//
+//  AwaitKit.h
+//  AwaitKit
+//
+//  Created by Artur Grigor on 26.01.2017.
+//  Copyright © 2017 Yannick Loriot. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for AwaitKit.
+FOUNDATION_EXPORT double AwaitKitVersionNumber;
+
+//! Project version string for AwaitKit.
+FOUNDATION_EXPORT const unsigned char AwaitKitVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <AwaitKit/PublicHeader.h>
+
+
